backend/requirements.txt
portia_sdk_python
fastapi>=0.68.0
uvicorn>=0.15.0